+++ CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Indicate status effects provided by custom Active Effects in the effects tabs
 - Add rich tooltips to Active Effects
 - Add scrolling text for updates to curse dice, armor, and injuries
+- Add injury dice and enhancement for Near Death level
 
 ### Changed